Initial Commit
Write a program that takes a list of unique strings as an argument, where each string is
a combination of a letter from a - z and a number from 0 - 99 , the initial character in
string being the letter. For example a23, d5, z0, q99 are some strings in this format. The
program should sort the list and return the lists L1 and L2 in the order mentioned below.
L1 : First list should contain all strings sorted in ascending order with respect to the first
character only. All the strings with the same initial character should be in the same order
as in the original list.
L2 : In the list L1 above, sort the strings starting with the same character, in descending
order with respect to the number formed by the remaining characters.
Sample Input 1
d34, g54, d12, b87, g1, c65, g40, g5, d77
Output
L1 : b87, c65, d34, d12, d77, g54, g1, g40, g5
L2 : b87, c65, d77, d34, d12, g54, g40, g5, g1